pyproject.toml
[tool.poetry]
name = "hotenov_com"
version = "2023.12.7"
description = "My personal website hotenov.com"
authors = ["Artem Hotenov <qa@hotenov.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2.8"
Pillow = "^10.1.0"
django-modeltranslation = "^0.18.11"
django-configurations = "^2.5"
gunicorn = "^21.2.0"
psycopg = {extras = ["binary"], version = "^3.1.14"}

[tool.poetry.group.test.dependencies]
pytest-django = "^4.7.0"
pytest-xdist = "*"
coverage = {extras = ["toml"], version = "^6.5.0"}
pytest-cov = "*"

[tool.poetry.group.lint.dependencies]
flake8 = "^5.0.4"
flake8-bandit = "^4.1.1"
flake8-bugbear = "^22.12.6"
flake8-docstrings = "^1.7.0"
flake8-black = "^0.3.6"
darglint = "^1.8.1"
mypy = "^0.971"

[tool.poetry.group.format.dependencies]
isort = "^5.12.0"
yapf = "^0.32.0"
black = "^22.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
force_single_line = true
force_sort_within_sections = false
lexicographical = true
order_by_type = false
group_by_package = true
src_paths = ["src", "tests"]
lines_after_imports = 2

[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

# pytest configuration
# Note: tool.pytest.ini_options will be replaced with tool.pytest in the future
[tool.pytest.ini_options]
# pytest-django config
DJANGO_SETTINGS_MODULE = "hotenov_com.settings"
DJANGO_CONFIGURATION = "TestSettings"
pythonpath = " . src"
python_files = "tests.py test_*.py *_tests.py"
markers = [
    "custom_domain(domain): marks tests to use custom domain for site",
]

# coverage configuration
[tool.coverage.run]
branch = true
source = ["src"]
omit = [
    "src/**/migrations/*",
]

[tool.coverage.paths]
source = [
    "src",
    ".nox/*/site-packages",
]

[tool.coverage.report]
show_missing = true
fail_under = 100
